+++ Application/DTOs/HomeWorkDTO/Queries/GetAllomeworkBounsViewModel.cs
@@ -8,8 +8,8 @@ namespace Application.DTOs
         public int Id { get; set; }
         public string Text { get; set; }
         public int MinCharacters { get; set; }
-        public int Points { get; set; }
-        public int BonusPoints { get; set; }
+        public double Points { get; set; }
+        public double BonusPoints { get; set; }
         public int BonusPointsStatus { get; set; }
         public int GroupInstanceId { get; set; }
         public string GroupInstanceSerial { get; set; }
